+++ example/lib/views/content/content_methods_view.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:synerise_flutter_sdk/model/content/screen_view_api_query.dart';
 import 'package:synerise_flutter_sdk/synerise.dart';
 
 import 'package:synerise_flutter_sdk_example/classes/utils.dart';
@@ -14,6 +15,7 @@ class _ContentMethodsViewState extends State<ContentMethodsView>
     with AutomaticKeepAliveClientMixin {
   final slugController = TextEditingController();
   final generateDocumentForm = GlobalKey<FormState>();
+  final generateDocumentWithApiQueryForm = GlobalKey<FormState>();
 
   final getRecommendationsV2Form = GlobalKey<FormState>();
   final slugRecoController = TextEditingController();
@@ -21,6 +23,7 @@ class _ContentMethodsViewState extends State<ContentMethodsView>
 
   final feedSlugController = TextEditingController();
   final generateScreenViewForm = GlobalKey<FormState>();
+  final generateScreenViewWithApiQueryForm = GlobalKey<FormState>();
 
   _tempFormBody() {
     return SingleChildScrollView(
@@ -49,7 +52,34 @@ class _ContentMethodsViewState extends State<ContentMethodsView>
                         onPressed: () =>
                             _generateDocumentCall(slugController.text),
                         icon: const Icon(Icons.file_copy_outlined),
-                        label: const Text('getDocument')),
+                        label: const Text('generateDocument')),
+                  ],
+                )),
+            const Divider(
+              thickness: 1,
+              color: Colors.grey,
+            ),
+            //generateDocument
+            Form(
+                key: generateDocumentWithApiQueryForm,
+                child: Column(
+                  children: [
+                    const Padding(
+                        padding: EdgeInsets.all(15),
+                        child: Text("generateDocumentWithApiQuery Test")),
+                    SizedBox(
+                        width: 350,
+                        child: TextFormField(
+                          controller: slugController,
+                          decoration: const InputDecoration(
+                              border: OutlineInputBorder(), labelText: "slug"),
+                          keyboardType: TextInputType.text,
+                        )),
+                    ElevatedButton.icon(
+                        onPressed: () =>
+                            _generateDocumentWithApiQueryCall(slugController.text),
+                        icon: const Icon(Icons.file_copy_outlined),
+                        label: const Text('generateDocumentWithApiQuery')),
                   ],
                 )),
             const Divider(
@@ -120,6 +150,29 @@ class _ContentMethodsViewState extends State<ContentMethodsView>
               thickness: 1,
               color: Colors.grey,
             ),
+                       Form(
+                key: generateScreenViewWithApiQueryForm,
+                child: Column(
+                  children: [
+                    const Padding(
+                        padding: EdgeInsets.all(15),
+                        child: Text("generateScreenViewWithApiQuery Test")),
+                    SizedBox(
+                        width: 350,
+                        child: TextFormField(
+                          controller: feedSlugController,
+                          decoration: const InputDecoration(
+                              border: OutlineInputBorder(),
+                              labelText: "feedSlug"),
+                          keyboardType: TextInputType.text,
+                        )),
+                    ElevatedButton.icon(
+                        onPressed: () =>
+                            _generateScreenViewWithApiQueryCall(feedSlugController.text),
+                        icon: const Icon(Icons.fit_screen),
+                        label: const Text('generateScreenView')),
+                  ],
+                )),
           ],
         ));
   }
@@ -128,6 +181,36 @@ class _ContentMethodsViewState extends State<ContentMethodsView>
     String slugName = slug;
 
     await Synerise.content.generateDocument(slugName,
+        onSuccess: (Document document) {
+      showDialog(
+          context: context,
+          builder: (context) {
+            return AlertDialog(
+                content: SingleChildScrollView(
+                    child: Column(children: [
+              const Text(
+                'Document Identifier',
+                style: TextStyle(fontStyle: FontStyle.italic),
+              ),
+              Container(
+                  width: double.infinity,
+                  margin: const EdgeInsets.all(5.0),
+                  padding: const EdgeInsets.all(5),
+                  decoration: BoxDecoration(
+                      border: Border.all(width: 0.5, color: Colors.black)),
+                  child: Text(document.identifier.toString(),
+                      textScaleFactor: 0.5))
+            ])));
+          });
+    }, onError: (SyneriseError error) {
+      Utils.displaySimpleAlert(error.message, context);
+    });
+  }
+
+  Future<void> _generateDocumentWithApiQueryCall(String slug) async {
+    DocumentApiQuery apiQuery = DocumentApiQuery(slug: slug);
+
+    await Synerise.content.generateDocumentWithApiQuery(apiQuery,
         onSuccess: (Document document) {
       showDialog(
           context: context,
@@ -191,6 +274,37 @@ class _ContentMethodsViewState extends State<ContentMethodsView>
   Future<void> _generateScreenViewCall(String feedSlug) async {
     String slug = feedSlug;
     await Synerise.content.generateScreenView(slug,
+        onSuccess: (ScreenView screenViewResponse) {
+      showDialog(
+          context: context,
+          builder: (context) {
+            return AlertDialog(
+                content: SingleChildScrollView(
+                    child: Column(children: [
+              const Text(
+                'ScreenView Data',
+                style: TextStyle(fontStyle: FontStyle.italic),
+              ),
+              Container(
+                  width: double.infinity,
+                  margin: const EdgeInsets.all(5.0),
+                  padding: const EdgeInsets.all(5),
+                  decoration: BoxDecoration(
+                      border: Border.all(width: 0.5, color: Colors.black)),
+                  child: Text(screenViewResponse.data.toString(),
+                      textScaleFactor: 0.5))
+            ])));
+          });
+    }, onError: (SyneriseError error) {
+      Utils.displaySimpleAlert(error.message, context);
+    });
+  }
+
+  Future<void> _generateScreenViewWithApiQueryCall(String feedSlug) async {
+    ScreenViewApiQuery apiQuery = ScreenViewApiQuery(feedSlug: feedSlug);
+    apiQuery.productId = "0000208094061-L";
+
+    await Synerise.content.generateScreenViewWithApiQuery(apiQuery,
         onSuccess: (ScreenView screenViewResponse) {
       showDialog(
           context: context,